Single repository. 1. Removed Repository interface 2. Renamed BaseRepository to Repository.
